spi: byte packing was the issue
If the transmission unit size is 8 bits or less and we access the data
register DR as a 16-bit value, the SPI peripheral assumes we have packed
two bytes in there and sends them both in sequence; similarly, reading
DR as a 16-bit value reads two bytes at once.

The stm32-rs team already solved this by providing 8-bit access to the
relevant register under the name dr8(). Accessing the register as an
8-bit value triggers the byte-for-byte behavior we want.